feat: implement bookings feature with CRUD operations and database integration

+++ src/features/calendars/data_transfer_objects.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize)]
 pub struct CalendarResponse {
-    pub id: u64,
+    pub id: u32,
     pub name: String,
 }
